refactor(course): split /courses/[slug]/page.tsx into _components
Move CourseHero (with progress meter and bookmark button) and LessonList
into the page-private _components folder.
